+++ app/random-id/CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Added
 
 - Added home page in which user can generate random id
-- Added options to control the alphabet of id
+- Added options to control alphabet and length of id